Menu is animated and the background is back in place
Plus, it doesn't get in the way of anything below it.

The background has been moved to the <ul> so that the menu items and
background are all one peice, and #menu never needs a height (which was
getting in the way even when it was off)

Also, another toggleProperty was removed and made more explicit. :-)