Signup/Signin is working, except for photo upload notification to rails
Added year to birthdate to fit with the existing user model
Validation for demographics is wonky and not built out yet. Still needs
to highlight problem fields, and should work after multiple errors
TOS confirmation is missing, should also be validated.
The user is created after the demographics step. Everything after that
is just updating the user, who is logged in at that point.
The photo upload form is not in an iframe for now, but probably should
be again.
Logout now directs to the landing page
The confirmation email works!